requirements.txt
Django==4.2.5
psycopg2>=2.8
django-debug-toolbar==4.2.0
djangorestframework~=3.14.0
django-filter==23.3
drf-yasg==1.21.7
drf-nested-routers==0.93.4
#pillow==10.0.1
django-admin-inline-paginator==0.4.0
sentry-sdk==1.31.0
django-ckeditor==6.7.0
django_admin_interface==0.26.1
whitenoise==6.5.0
django-redis==5.4.0
django-cors-headers==4.2.0
django-import-export==3.3.1
social-auth-app-django==5.3.0
djangorestframework-simplejwt==5.3.0
django-rest-passwordreset==1.3.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
django-unicorn==0.61.0